FormGroupHiddenProp
 * Functions passed as values in `inputProperties` are now called with an additional parameter: `document`
 * `FormGroup` now accepts a `hidden` property, which can also be a function
 * Split bootstrap-specific code out to `FormGroupDefault`
 * Implemented `group.hidden` prop in `vulcan:ui-bootstrap` and `vulcan:ui-material`
