+++ packages/vulcan-forms/lib/components/FormGroup.jsx
@@ -1,53 +1,9 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { Components, instantiateComponent, Utils } from 'meteor/vulcan:core';
-import classNames from 'classnames';
+import { instantiateComponent } from 'meteor/vulcan:core';
 import { registerComponent, mergeWithComponents } from 'meteor/vulcan:core';
 import Users from 'meteor/vulcan:users';
 
-const FormGroupHeader = ({ toggle, collapsed, label }) => (
-  <div className="form-section-heading" onClick={toggle}>
-    <h3 className="form-section-heading-title">{label}</h3>
-    <span className="form-section-heading-toggle">
-      {collapsed ? (
-        <Components.IconRight height={16} width={16}/>
-      ) : (
-        <Components.IconDown height={16} width={16}/>
-      )}
-    </span>
-  </div>
-);
-FormGroupHeader.propTypes = {
-  toggle: PropTypes.func.isRequired,
-  label: PropTypes.string.isRequired,
-  collapsed: PropTypes.bool,
-  group: PropTypes.object,
-};
-registerComponent({ name: 'FormGroupHeader', component: FormGroupHeader });
-
-const FormGroupLayout = ({ children, label, anchorName, heading, collapsed, group, hasErrors }) => (
-  <div className={`form-section form-section-${anchorName} form-section-${Utils.slugify(label)}`}>
-    <a name={anchorName}/>
-    {heading}
-    <div
-      className={classNames({
-        'form-section-collapsed': collapsed && !hasErrors
-      })}
-    >
-      {children}
-    </div>
-  </div>
-);
-FormGroupLayout.propTypes = {
-  label: PropTypes.string,
-  anchorName: PropTypes.string,
-  heading: PropTypes.node,
-  collapsed: PropTypes.bool,
-  group: PropTypes.object,
-  hasErrors: PropTypes.bool,
-  children: PropTypes.node
-};
-registerComponent({ name: 'FormGroupLayout', component: FormGroupLayout });
 
 class FormGroup extends PureComponent {
   constructor (props) {
@@ -71,6 +27,7 @@ class FormGroup extends PureComponent {
         toggle={this.toggle}
         label={this.props.label}
         collapsed={this.state.collapsed}
+        hidden={this.isHidden()}
         group={this.props.group}
       />
     );
@@ -83,6 +40,14 @@ class FormGroup extends PureComponent {
         .length;
     });
   
+  isHidden = () => {
+    const { hidden, document } = this.props;
+    const isHidden = typeof hidden === 'function'
+      ? hidden({ ...this.props, document })
+      : hidden || false;
+    return isHidden;
+  };
+  
   render () {
     if (this.props.group.adminsOnly && !Users.isAdmin(this.props.currentUser)) {
       return null;
@@ -90,7 +55,7 @@ class FormGroup extends PureComponent {
   
     const { name, fields, formComponents, label, group } = this.props;
     const { collapsed } = this.state;
-    
+  
     const FormComponents = mergeWithComponents(formComponents);
     const anchorName = name.split('.').length > 1 ? name.split('.')[1] : name;
     
@@ -100,6 +65,7 @@ class FormGroup extends PureComponent {
           anchorName={anchorName}
           toggle={this.toggle}
           collapsed={collapsed}
+          hidden={this.isHidden()}
           group={group}
           heading={name === 'default' ? null : this.renderHeading(FormComponents)}
           hasErrors={this.hasErrors()}
@@ -137,6 +103,7 @@ FormGroup.propTypes = {
   name: PropTypes.string,
   label: PropTypes.string,
   order: PropTypes.number,
+  hidden: PropTypes.func,
   fields: PropTypes.array.isRequired,
   group: PropTypes.object.isRequired,
   errors: PropTypes.array.isRequired,
@@ -154,47 +121,3 @@ FormGroup.propTypes = {
 export default FormGroup;
 
 registerComponent('FormGroup', FormGroup);
-
-const IconRight = ({ width = 24, height = 24 }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={width}
-    height={height}
-    viewBox="0 0 24 24"
-  >
-    <polyline
-      fill="none"
-      stroke="#000"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeMiterlimit="10"
-      points="5.5,23.5 18.5,12 5.5,0.5"
-      id="Outline_Icons"
-    />
-    <rect fill="none" width="24" height="24" id="Frames-24px"/>
-  </svg>
-);
-
-registerComponent('IconRight', IconRight);
-
-const IconDown = ({ width = 24, height = 24 }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={width}
-    height={height}
-    viewBox="0 0 24 24"
-  >
-    <polyline
-      fill="none"
-      stroke="#000"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeMiterlimit="10"
-      points="0.501,5.5 12.001,18.5 23.501,5.5"
-      id="Outline_Icons"
-    />
-    <rect fill="none" width="24" height="24" id="Frames-24px"/>
-  </svg>
-);
-
-registerComponent('IconDown', IconDown);
