+++ packages/vulcan-ui-bootstrap/lib/modules/components.js
@@ -1,4 +1,3 @@
-import '../components/forms/FormElement.jsx';
 import '../components/forms/Checkbox.jsx';
 import '../components/forms/Checkboxgroup.jsx';
 import '../components/forms/Date.jsx';
@@ -6,17 +5,19 @@ import '../components/forms/Date2.jsx';
 import '../components/forms/Datetime.jsx';
 import '../components/forms/Default.jsx';
 import '../components/forms/Email.jsx';
+import '../components/forms/FormComponentInner.jsx';
+import '../components/forms/FormControl.jsx'; // note: only used by old accounts package, remove soon?
+import '../components/forms/FormElement.jsx';
+import '../components/forms/FormGroupDefault';
+import '../components/forms/FormItem.jsx';
 import '../components/forms/Number.jsx';
 import '../components/forms/Radiogroup.jsx';
 import '../components/forms/Select.jsx';
 import '../components/forms/SelectMultiple.jsx';
+import '../components/forms/StaticText.jsx';
 import '../components/forms/Textarea.jsx';
 import '../components/forms/Time.jsx';
 import '../components/forms/Url.jsx';
-import '../components/forms/StaticText.jsx';
-import '../components/forms/FormComponentInner.jsx';
-import '../components/forms/FormControl.jsx'; // note: only used by old accounts package, remove soon?
-import '../components/forms/FormItem.jsx';
 
 import '../components/ui/Button.jsx';
 import '../components/ui/Alert.jsx';
