+++ packages/vulcan-ui-bootstrap/lib/stylesheets/style.scss
@@ -156,6 +156,10 @@ div.ReactTags__suggestions mark{
   display: none;
 }
 
+.form-section-hidden {
+  display: none;
+}
+
 .form-errors{
   .alert{
     ul{
